v534 поддержка номеров до 8 байт

+++ sources/Reader/src/Nodes/OSDP/OSDP.h
@@ -58,6 +58,6 @@ namespace OSDP
 
     namespace Card
     {
-        void Insert(uint number);
+        void Insert(uint64 number);
     }
 }
